# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Twindix Admin API",
        "description": "API documentation for Twindix Admin Panel - Access Control System",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/api/user-types": {
            "get": {
                "tags": [
                    "Types"
                ],
                "summary": "List types with subtypes",
                "operationId": "95026e7d7a2e15abefa57c684af7896b",
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Types"
                ],
                "summary": "Create type",
                "operationId": "dbc16e3fad5efdbcae92b5e9b44cbd0d",
                "responses": {
                    "201": {
                        "description": "Created"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/user-types/{id}": {
            "put": {
                "tags": [
                    "Types"
                ],
                "summary": "Update type",
                "operationId": "ed16d0dd6b5c897d6d8a3f873508e2e1",
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Types"
                ],
                "summary": "Delete type",
                "operationId": "0816ce731ead5c6b222b7815b923e5ad",
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/user-types/{id}/sub-types": {
            "post": {
                "tags": [
                    "Types"
                ],
                "summary": "Create subtype",
                "operationId": "034a2d3754fe5b8d4c6ade99e2affd2b",
                "responses": {
                    "201": {
                        "description": "Created"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/user-sub-types/{id}": {
            "put": {
                "tags": [
                    "Types"
                ],
                "summary": "Update subtype",
                "operationId": "deca8ccb69539f7525be443528f57e53",
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Types"
                ],
                "summary": "Delete subtype",
                "operationId": "a356aad7f7190762cbefa67e06eeeb26",
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}